Initial file upload
Initial upload of files for community project.
Current Functionality:
Create Account
Login
Logout

When you download the files, add everything to any local folder on your
computer except for the mysql folder. The mysql folder contains the sql
code you’ll need to create a test table for “users” in your local mysql
database.

After this download the node_modules folder I will email you and add it
to the same folder all your other files are in. To run the
program/start the node server, type “node server start” into the
command line/terminal but first cd to the folder the project is stored
in.