Completed o_events.hjs and others
Finished the o_events.hjs view. Also modified bits of the other views
so everything could look standardized. Modified some of the queries to
be more specific and fixed the error for adding a volunteer to an
event. If you’re having trouble with the volunteering functionality,
the problem will most likely be on your database. The only thing I can
think of is clearing your tables and creating them again, or creating
another test db.