Added upload images functionality
Added ability to upload images for events. Events with an uploaded
image now display that image, and events without one just display the
default image. Also added the new navbar to all views except for the
login view. Modified the database as well so update with the
mysql/images.sql file